feat(composer): add modulation, fortspinnung sequences, and invertible counterpoint
Extend the composer subsystem with three harmonic and contrapuntal
capabilities, each wired through CandidateSearch provenance bits, the
Validator, and the HarnessFixture layouts.

Modulation / tonicization / borrowed chords:
  - Add ModulationType (Pivot/Phrase/CommonTone) and ModulationEvent to
    HarmonicPlan; tag ChordEvent with secondary_of, is_borrowed, and
    is_picardy.
  - applyP8Bits sets ModulationCommitted, SecondaryDominantResolved,
    PicardyThird, and ModalMixture; CandidateSearch biases the Picardy
    major third in final tonic chords.
  - Validator adds modulation_pivot_chord_required and
    secondary_dominant_resolution, and exempts secondary-dominant chord
    tones from augmented_melodic, tritone_melodic, and diminished_melodic.

Fortspinnung sequences and imitation:
  - Add SequencePattern, SequenceTemplate, and ImitationEntry to Material
    and the FortspinnungSpan VoiceIntent; CandidateSearch replays seed
    motifs transposed by descending-fifths, descending-step, or
    ascending-step patterns, setting FortspinnungSourced, SequenceStep,
    and ImitationEntryMatched.

Invertible counterpoint at the octave:
  - Add InvertibleAt8va provenance bit with a strong-beat perfect-fourth
    pre-filter scoped to upper-adjacent voice pairs; Validator adds
    invertible_at_octave and fourth_only_on_weak_beat.

Emit RuleIdMask as a full 64-bit unsigned integer in json_export to
avoid truncating rule bits >= 32; add JsonWriter::value(uint64_t).

Wire Phase8/Phase9/Phase10 fixtures, CLI parsing, and closure tags;
add candidate_search, json_export, and validator tests.

+++ src/composer/json_export.cpp
@@ -85,8 +85,11 @@ std::string emitProvenanceJson(const std::vector<NoteProvenance>& provenance) {
     w.key("candidate_score");
     w.value(static_cast<double>(p.candidate_score));
     w.key("satisfied_rules");
-    // Emit as integer; bach-mcp parses it as a bitset on its side.
-    w.value(static_cast<int>(p.satisfied_rules));
+    // Emit as a full 64-bit unsigned integer; bach-mcp parses it as a
+    // bitset. A 32-bit cast would truncate rule bits >= 32 (P9/P10) and,
+    // worse, make bit 31 the sign bit so downstream masking of high bits
+    // sign-extends — so the wide RuleIdMask must be written verbatim.
+    w.value(static_cast<std::uint64_t>(p.satisfied_rules));
     w.key("rejected_alternatives");
     w.value(static_cast<int>(p.rejected_alternatives));
     w.endObject();
